Refactor after major navbar change
